renamed btc to bitcoin, removed klayton, and moved evm/utils functions to inbounds.go
