+++ zetaclient/bitcoin/bitcoin_signer.go
