+++ zetaclient/bitcoin/bitcoin_signer_test.go
